fix : 라우팅 관련 경로 수정 - App.jsx 라우팅 경로 추가 - Posting.jsx로 UI 구현 - Landing 페이지는 일단 링크만 연결해둠 ref : #5
