--- src/pages/Recruitment/Board.jsx
@@ -1,1 +1,0 @@
-// 온보딩 화면
